% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{Wind radii charts}
\usage{
geom_hurricane(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, scale_radii = 0.1, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes or aes_.
If specified and inherit.aes = TRUE (the default), it is combined
with the default mapping at the top level of the plot. You must
supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot.

A data.frame, or other object, will override the plot data. All objects will be fortified to produce
a data frame. See fortify for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame.,
and will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning.
If TRUE, missing values are silently removed}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.
This is most useful for helper functions that define both data and aesthetics and shouldn't inherit
behaviour from the default plot specification, e.g. borders.}

\item{scale_radii}{A numeric of scale factor for wind radii chart, default value is 10%.}

\item{...}{other arguments passed on to layer. These are often aesthetics, used to set an aesthetic to
a fixed value, like color = "red" or size = 3. They may also be parameters to the paired geom/stat.}
}
\description{
This wind radii chart displays how far winds of a certain intensity (e.g., 34, 50, or 64 knots)
extended from a hurricane’s center, with separate values given for the northeast, northwest,
southeast, and southwest quadrants of the storm.
}
\section{Aesthetics}{

\aesthetics{geom}{hurricane}
}

\examples{
\dontrun {
# hurrican data of one observation
library(ggmap)
library(ggplot2)
ike_2008_09_13 <- hurricanes[hurricanes$storm_id=="IKE-2008" &
hurricanes$date_time=="2008-09-13 12:00:00", ]
# Get base map near Houston area where hurrican IKE passed by
map_data <- get_map("Houston", zoom = 6, maptype = "toner-background")
# Plot the hurricane
base_map <- ggmap(map_data, extent = "device")
base_map +
geom_hurricane(data = ike_2008_09_13, aes(x = longitude, y = latitude,
r_ne = ne, r_se = se, r_nw = nw, r_sw = sw, fill = wind_speed2,
color = wind_speed2), scale_radii = 0.2) +
scale_color_manual(name = "Wind speed (kts)",
                   values = c("red", "orange", "yellow")) +
scale_fill_manual(name = "Wind speed (kts)",
                   values = c("red", "orange", "yellow"))
}
}
